% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_geoarrow_array}
\alias{as_geoarrow_array}
\alias{as_geoarrow_array_stream}
\title{Convert an object to a GeoArrow array}
\usage{
as_geoarrow_array(x, ..., schema = NULL)

as_geoarrow_array_stream(x, ..., schema = NULL)
}
\arguments{
\item{x}{An object}

\item{...}{Passed to S3 methods}

\item{schema}{A geoarrow extension schema to use as the target type}
}
\value{
A \link[nanoarrow:as_nanoarrow_array]{nanoarrow_array}.
}
\description{
Convert an object to a GeoArrow array
}
\examples{
as_geoarrow_array(wk::wkt("POINT (0 1)"))

}
