% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{as_geoarrow_table}
\alias{as_geoarrow_table}
\title{Create Arrow Tables}
\usage{
as_geoarrow_table(
  handleable,
  schema = NULL,
  strict = FALSE,
  null_point_as_empty = FALSE,
  geoparquet_metadata = FALSE
)
}
\arguments{
\item{handleable}{An object with a \code{\link[wk:wk_handle]{wk::wk_handle()}} method}

\item{schema}{A \code{\link[narrow:narrow_schema]{narrow::narrow_schema()}} to use as a storage method.}

\item{strict}{Use \code{TRUE} to respect choices of storage type, dimensions,
and CRS provided by \code{schema}. The default, \code{FALSE}, updates these values
to match the data.}

\item{null_point_as_empty}{Use \code{TRUE} when creating point arrays for Parquet
to work around a bug whereby null items cannot be re-read by Arrow Parquet
readers.}

\item{geoparquet_metadata}{Use \code{TRUE} to add GeoParquet metadata to the
output schema metadata.}
}
\value{
an \link[arrow:Table]{arrow::Table}.
}
\description{
Like \code{\link[arrow:as_arrow_table]{arrow::as_arrow_table()}}, \code{as_geoarrow_table()} creates an
\link[arrow:Table]{arrow::Table} object. The geo-enabled version has a few additional options
that customize the specific output schema and table-level metadata
that is generated.
}
\examples{
as_geoarrow_table(data.frame(col1 = 1:5, col2 = wk::xy(1:5, 6:10)))

}
