% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{geo_arrow_metadata}
\alias{geo_arrow_metadata}
\title{Extract Arrow extension type  metadata}
\usage{
geo_arrow_metadata(schema)
}
\arguments{
\item{schema}{A \code{\link[=carrow_schema]{carrow_schema()}}}
}
\value{
A \code{list()}
}
\description{
Extract Arrow extension type  metadata
}
\examples{
geo_arrow_metadata(geo_arrow_schema_point(crs = "OGC:CRS84"))

}
