% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoarrow-package.R
\docType{package}
\name{geoarrow-package}
\alias{geoarrow}
\alias{geoarrow-package}
\title{geoarrow: Extension Types for Spatial Data for Use with 'Arrow'}
\description{
Provides extension types and conversions to between R-native object types and 'Arrow' columnar types. This includes integration among the 'arrow', 'nanoarrow', 'sf', and 'wk' packages such that spatial metadata is preserved wherever possible. Extension type implementations ensure first-class geometry data type support in the 'arrow' and 'nanoarrow' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://geoarrow.org/geoarrow-r/}
  \item \url{https://github.com/geoarrow/geoarrow-r}
  \item Report bugs at \url{https://github.com/geoarrow/geoarrow-r/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})
}

Other contributors:
\itemize{
  \item Anthony North \email{anthony.jl.north@gmail.com} [contributor]
  \item Apache Software Foundation \email{dev@arrow.apache.org} [copyright holder]
  \item Ulf Adams [copyright holder]
  \item Daniel Lemire [copyright holder]
  \item Joao Paulo Magalhaes [copyright holder]
}

}
\keyword{internal}
