% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{geoarrow}
\alias{geoarrow}
\alias{geoarrow_wkb}
\alias{geoarrow_wkt}
\alias{as_geoarrow}
\alias{as_geoarrow.geoarrow_vctr}
\title{Create GeoArrow Vectors}
\usage{
geoarrow(x = wk::wkb(crs = NULL))

geoarrow_wkb(x = wk::wkb(crs = NULL))

geoarrow_wkt(x = wk::wkb(crs = NULL))

as_geoarrow(x, ..., ptype = NULL)

\method{as_geoarrow}{geoarrow_vctr}(x, ..., ptype = NULL)
}
\arguments{
\item{x}{An object that implements \code{\link[wk:wk_handle]{wk::wk_handle()}}}

\item{...}{Passed to S3 methods}

\item{ptype}{A prototype for the final vctr type, or NULL to guess the most
appropriate type.}
}
\value{
A vector class inheriting from 'geoarrow_vctr'.
}
\description{
The 'geoarrow_vctr' family of vector classes are a zero-copy wrapper around
an Arrow C Data interface representation of a \link[arrow:ChunkedArray]{arrow::ChunkedArray}. This
class provides most of the features that you might want when loading a file
(e.g., subset, print, format, inspect), and can be converted to some other
vector class (e.g., using \code{\link[sf:st_as_sf]{sf::st_as_sf()}} or \code{\link[=geoarrow_collect_sf]{geoarrow_collect_sf()}}) with
more features when appropriate.
}
