% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{geoarrow_create}
\alias{geoarrow_create}
\title{Create GeoArrow vectors}
\usage{
geoarrow_create(x, ..., schema = geoarrow_schema_default(x), strict = FALSE)
}
\arguments{
\item{x}{An object to convert to a GeoArrow representation.}

\item{...}{Passed to \code{\link[wk:wk_handle]{wk::wk_handle()}}}

\item{schema}{A \code{\link[narrow:narrow_schema]{narrow::narrow_schema()}} to use as a storage method.}

\item{strict}{Use \code{TRUE} to respect choices of storage type, dimensions,
and CRS provided by \code{schema}. The default, \code{FALSE}, updates these values
to match the data.}
}
\value{
An object of class 'narrow_vctr_geoarrow_EXTENSION' and
\code{\link[=narrow_vctr]{narrow_vctr()}}.
}
\description{
Create GeoArrow vectors
}
