% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{geoarrow_create_narrow}
\alias{geoarrow_create_narrow}
\alias{geoarrow_create_narrow_from_buffers}
\alias{geoarrow_schema_default}
\title{Create GeoArrow Arrays}
\usage{
geoarrow_create_narrow(
  handleable,
  ...,
  schema = geoarrow_schema_default(handleable),
  strict = FALSE,
  null_point_as_empty = FALSE
)

geoarrow_create_narrow_from_buffers(
  handleable,
  ...,
  schema = geoarrow_schema_default(handleable),
  strict = FALSE,
  null_point_as_empty = FALSE
)

geoarrow_schema_default(handleable, point = geoarrow_schema_point())
}
\arguments{
\item{handleable}{An object with a \code{\link[wk:wk_handle]{wk::wk_handle()}} method}

\item{...}{Passed to \code{\link[wk:wk_handle]{wk::wk_handle()}}}

\item{schema}{A \code{\link[narrow:narrow_schema]{narrow::narrow_schema()}} to use as a storage method.}

\item{strict}{Use \code{TRUE} to respect choices of storage type, dimensions,
and CRS provided by \code{schema}. The default, \code{FALSE}, updates these values
to match the data.}

\item{null_point_as_empty}{Use \code{TRUE} when creating point arrays for Parquet
to work around a bug whereby null items cannot be re-read by Arrow Parquet
readers.}

\item{point}{The point schema to use for coordinates}
}
\value{
A \code{\link[narrow:narrow_array]{narrow::narrow_array()}}
}
\description{
Create GeoArrow Arrays
}
\examples{
geoarrow_create_narrow_from_buffers(wk::xy(1:5, 1:5))

}
