% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{geoarrow_example}
\alias{geoarrow_example}
\alias{geoarrow_example_Array}
\alias{geoarrow_example_narrow_array}
\title{Create example geometry objects}
\usage{
geoarrow_example(
  which = "nc",
  schema = NULL,
  strict = FALSE,
  crs = NULL,
  edges = "planar"
)

geoarrow_example_Array(
  which = "nc",
  schema = NULL,
  strict = FALSE,
  crs = NULL,
  edges = "planar"
)

geoarrow_example_narrow_array(
  which = "nc",
  schema = NULL,
  strict = FALSE,
  crs = NULL,
  edges = "planar"
)
}
\arguments{
\item{which}{An example name. Valid example names are
\itemize{
\item "nc"
\item "point", "linestring", "polygon", "multipoint",
"multilinestring", "multipolygon", "geometrycollection"
\item One of the above with the "_z", "_m", or "_zm" prefix.
}}

\item{schema}{A \code{\link[narrow:narrow_schema]{narrow::narrow_schema()}} to use as a storage method.}

\item{strict}{Use \code{TRUE} to respect choices of storage type, dimensions,
and CRS provided by \code{schema}. The default, \code{FALSE}, updates these values
to match the data.}

\item{crs}{A length-one character representation of the CRS. The WKT2
representation is recommended as the most complete way to encode this
information; however, any string that can be recognized by the PROJ
command-line utility (e.g., "OGC:CRS84").}

\item{edges}{Use "spherical" to assert that edges should be interpolated
using the shortest geodesic path (great circle on a sphere).}
}
\value{
The result of \code{\link[=geoarrow_create]{geoarrow_create()}} with the specified example type.
}
\description{
Create example geometry objects
}
\examples{
geoarrow_example("polygon")

}
