% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{geoarrow_metadata}
\alias{geoarrow_metadata}
\title{Extract Arrow extension type  metadata}
\usage{
geoarrow_metadata(schema)
}
\arguments{
\item{schema}{A \code{\link[=narrow_schema]{narrow_schema()}}}
}
\value{
A \code{list()}
}
\description{
Extract Arrow extension type  metadata
}
\examples{
geoarrow_metadata(geoarrow_schema_point(crs = "OGC:CRS84"))

}
