% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{geoarrow_schema_parse}
\alias{geoarrow_schema_parse}
\alias{is_geoarrow_schema}
\alias{as_geoarrow_schema}
\title{Inspect a GeoArrow schema}
\usage{
geoarrow_schema_parse(
  schema,
  extension_name = NULL,
  infer_from_storage = FALSE
)

is_geoarrow_schema(schema)

as_geoarrow_schema(schema)
}
\arguments{
\item{schema}{A \link[nanoarrow:as_nanoarrow_schema]{nanoarrow_schema}}

\item{extension_name}{An extension name to use if schema is a storage type.}

\item{infer_from_storage}{Attempt to guess an extension name if schema is not
a geoarrow extension type.}
}
\value{
A list of parsed properties
}
\description{
Inspect a GeoArrow schema
}
\examples{
geoarrow_schema_parse(na_extension_geoarrow("POINT"))

}
