% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{geoarrow_schema_point}
\alias{geoarrow_schema_point}
\alias{geoarrow_schema_point_struct}
\alias{geoarrow_schema_linestring}
\alias{geoarrow_schema_polygon}
\alias{geoarrow_schema_multi}
\alias{geoarrow_schema_sparse_geometrycollection}
\alias{geoarrow_schema_dense_geometrycollection}
\alias{geoarrow_schema_wkb}
\alias{geoarrow_schema_wkt}
\alias{geoarrow_schema_geojson}
\title{Create low-level Arrow schemas}
\usage{
geoarrow_schema_point(
  name = "",
  dim = "xy",
  crs = NULL,
  format_coord = "g",
  nullable = TRUE
)

geoarrow_schema_point_struct(
  name = "",
  dim = "xy",
  crs = NULL,
  format_coord = "g",
  nullable = TRUE
)

geoarrow_schema_linestring(
  name = "",
  format = "+l",
  nullable = TRUE,
  geodesic = FALSE,
  point = geoarrow_schema_point(nullable = FALSE)
)

geoarrow_schema_polygon(
  name = "",
  format = c("+l", "+l"),
  nullable = TRUE,
  geodesic = FALSE,
  point = geoarrow_schema_point(nullable = FALSE)
)

geoarrow_schema_multi(child, name = "", format = "+l", nullable = TRUE)

geoarrow_schema_sparse_geometrycollection(children = list(), name = "")

geoarrow_schema_dense_geometrycollection(children = list(), name = "")

geoarrow_schema_wkb(
  name = "",
  format = "z",
  crs = NULL,
  geodesic = FALSE,
  nullable = TRUE
)

geoarrow_schema_wkt(
  name = "",
  format = "u",
  crs = NULL,
  geodesic = FALSE,
  nullable = TRUE
)

geoarrow_schema_geojson(name = "", format = "u", crs = NULL, nullable = TRUE)
}
\arguments{
\item{name}{An optional name}

\item{dim}{A string with one character per dimension. The string must be one
of xy, xyz, xym, or xyzm.}

\item{crs}{A length-one character representation of the CRS. The WKT2
representation is recommended as the most complete way to encode this
information; however, any string that can be recognized by the PROJ
command-line utility (e.g., "OGC:CRS84").}

\item{format_coord}{A format for floating point coordinate storage. This
can be "f" (float/float32) or "g" (double/float64).}

\item{nullable}{\code{TRUE} if the type is semantically nullable}

\item{format}{For list-of types, the storage format. This can be +l
(32-bit integer offsets), +L (64-bit integer offsets), or +w:(fixed_size).
For \code{\link[=geoarrow_schema_polygon]{geoarrow_schema_polygon()}}, \code{format} has two elements: the first
for the list of rings and the second for the list of points.}

\item{geodesic}{Use \code{TRUE} to assert that edges should be interpolated
using the shortest geodesic path (great circle on a sphere).}

\item{point}{The point schema to use for coordinates}

\item{child}{The child schema to use in a single-type (multi) collection}

\item{children}{The child schemas for children in the union}
}
\value{
A \code{\link[=sparrow_schema]{sparrow_schema()}}.
}
\description{
These schemas are used as the basis for column types in Apache Arrow
}
\examples{
geoarrow_schema_point()
geoarrow_schema_linestring()
geoarrow_schema_polygon()
geoarrow_schema_multi(geoarrow_schema_point())
geoarrow_schema_sparse_geometrycollection()
geoarrow_schema_dense_geometrycollection()

}
