% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-default.R
\name{infer_geoarrow_schema}
\alias{infer_geoarrow_schema}
\title{Infer a GeoArrow-native type from a vector}
\usage{
infer_geoarrow_schema(x, ..., promote_multi = TRUE, coord_type = NULL)
}
\arguments{
\item{x}{An object from which to infer a schema.}

\item{...}{Passed to S3 methods.}

\item{promote_multi}{Use \code{TRUE} to return a MULTI type when both normal and
MULTI elements are in the same array.}

\item{coord_type}{Specify the coordinate type to use if returning}
}
\value{
A \link[=as_nanoarrow_schema]{nanoarrow_schema}
}
\description{
Infer a GeoArrow-native type from a vector
}
\examples{
infer_geoarrow_schema(wk::wkt("POINT (0 1)"))

}
