% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-sf.R, R/read.R
\name{read_geoparquet_sf}
\alias{read_geoparquet_sf}
\alias{geoarrow_collect_sf}
\alias{read_geoparquet}
\alias{geoarrow_collect}
\alias{geoarrow_collect.data.frame}
\alias{geoarrow_collect.Table}
\alias{geoarrow_collect.RecordBatch}
\alias{geoarrow_collect.RecordBatchReader}
\alias{geoarrow_collect.Dataset}
\alias{geoarrow_collect.arrow_dplyr_query}
\title{Read geometry from Apache Parquet files}
\usage{
read_geoparquet_sf(file, ...)

geoarrow_collect_sf(x, ..., metadata = NULL)

read_geoparquet(
  file,
  ...,
  as_data_frame = TRUE,
  handler = NULL,
  metadata = NULL
)

geoarrow_collect(x, ..., handler = NULL, metadata = NULL)

\method{geoarrow_collect}{data.frame}(x, ..., handler = NULL, metadata = NULL)

\method{geoarrow_collect}{Table}(x, ..., handler = NULL, metadata = NULL)

\method{geoarrow_collect}{RecordBatch}(x, ..., handler = NULL, metadata = NULL)

\method{geoarrow_collect}{RecordBatchReader}(x, trans = identity, ..., handler = NULL, metadata = NULL)

\method{geoarrow_collect}{Dataset}(x, trans = identity, ..., handler = NULL, metadata = NULL)

\method{geoarrow_collect}{arrow_dplyr_query}(x, ..., handler = NULL, metadata = NULL)
}
\arguments{
\item{file}{A file or InputStream to read; passed to
\code{\link[arrow:read_parquet]{arrow::read_parquet()}}, \code{\link[arrow:read_feather]{arrow::read_feather()}}, or
\code{\link[arrow:read_ipc_stream]{arrow::read_ipc_stream()}}.}

\item{...}{
  Arguments passed on to \code{\link[arrow:read_parquet]{arrow::read_parquet}}
  \describe{
    \item{\code{col_select}}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:vars_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}
    \item{\code{props}}{\link[arrow]{ParquetArrowReaderProperties}}
  }}

\item{x}{An object to collect into a data.frame, converting geometry
columns according to \code{handler}.}

\item{metadata}{Optional metadata to include to override metadata available
in the file.}

\item{as_data_frame}{Use \code{FALSE} to return an \link[arrow:Table]{arrow::Table}
instead of a data.frame.}

\item{handler}{A \link[wk:wk_handle]{wk handler} to use when \code{as.data.frame}
is TRUE for all geometry columns.}

\item{trans}{A function to be applied to each chunk after it has been
collected into a data frame.}
}
\value{
The result of \code{\link[arrow:read_parquet]{arrow::read_parquet()}}, with geometry
columns processed according to \code{handler}.
}
\description{
Read geometry from Apache Parquet files
}
