% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{as_geoarrow_vctr}
\alias{as_geoarrow_vctr}
\title{GeoArrow encoded arrays as R vectors}
\usage{
as_geoarrow_vctr(x, ..., schema = NULL)
}
\arguments{
\item{x}{An object that works with \code{\link[=as_geoarrow_array_stream]{as_geoarrow_array_stream()}}. Most
spatial objects in R already work with this method.}

\item{...}{Passed to \code{\link[=as_geoarrow_array_stream]{as_geoarrow_array_stream()}}}

\item{schema}{An optional \code{schema} (e.g., \code{\link[=na_extension_geoarrow]{na_extension_geoarrow()}}).}
}
\value{
A vctr of class 'geoarrow_vctr'
}
\description{
GeoArrow encoded arrays as R vectors
}
\examples{
as_geoarrow_vctr("POINT (0 1)")

}
