% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{geo_arrow_schema_point}
\alias{geo_arrow_schema_point}
\alias{geo_arrow_schema_linestring}
\alias{geo_arrow_schema_polygon}
\alias{geo_arrow_schema_multi}
\alias{geo_arrow_schema_sparse_geometrycollection}
\alias{geo_arrow_schema_dense_geometrycollection}
\alias{geo_arrow_schema_wkb}
\alias{geo_arrow_schema_wkt}
\alias{geo_arrow_schema_geojson}
\alias{geo_arrow_schema_large_geojson}
\alias{geo_arrow_schema_point_struct_float64}
\alias{geo_arrow_schema_point_struct_float32}
\alias{geo_arrow_schema_point_float64}
\alias{geo_arrow_schema_point_float32}
\alias{geo_arrow_schema_point_s2}
\alias{geo_arrow_schema_point_h3}
\title{Create low-level Arrow schemas}
\usage{
geo_arrow_schema_point(name = NULL, dim = "xy", crs = NULL, nullable = TRUE)

geo_arrow_schema_linestring(
  name = NULL,
  format = "+l",
  nullable = TRUE,
  ellipsoidal = FALSE,
  point = geo_arrow_schema_point(nullable = FALSE)
)

geo_arrow_schema_polygon(
  name = NULL,
  format = c("+l", "+l"),
  nullable = TRUE,
  ellipsoidal = FALSE,
  point = geo_arrow_schema_point(nullable = FALSE)
)

geo_arrow_schema_multi(child, format = "+l", name = NULL, nullable = TRUE)

geo_arrow_schema_sparse_geometrycollection(children = list(), name = NULL)

geo_arrow_schema_dense_geometrycollection(children = list(), name = NULL)

geo_arrow_schema_wkb(
  name = NULL,
  format = "z",
  crs = NULL,
  ellipsoidal = FALSE,
  nullable = TRUE
)

geo_arrow_schema_wkt(
  name = NULL,
  format = "u",
  crs = NULL,
  ellipsoidal = FALSE,
  nullable = TRUE
)

geo_arrow_schema_geojson(
  name = NULL,
  format = "u",
  crs = NULL,
  nullable = TRUE
)

geo_arrow_schema_large_geojson(name = NULL, crs = NULL, nullable = TRUE)

geo_arrow_schema_point_struct_float64(
  name = NULL,
  dim = "xy",
  crs = NULL,
  nullable = TRUE
)

geo_arrow_schema_point_struct_float32(
  name = NULL,
  dim = "xy",
  crs = NULL,
  nullable = TRUE
)

geo_arrow_schema_point_float64(
  name = NULL,
  dim = "xy",
  crs = NULL,
  nullable = TRUE
)

geo_arrow_schema_point_float32(
  name = NULL,
  dim = "xy",
  crs = NULL,
  nullable = TRUE
)

geo_arrow_schema_point_s2(name = NULL, crs = "OGC:CRS84", nullable = TRUE)

geo_arrow_schema_point_h3(name = NULL, crs = "OGC:CRS84", nullable = TRUE)
}
\arguments{
\item{crs}{A length-one character representation of the CRS. The WKT2
representation is recommended as the most complete way to encode this
information; however, any string that can be recognized by the PROJ
command-line utility (e.g., "OGC:CRS84").}

\item{ellipsoidal}{Use \code{TRUE} to assert that edges should be interpolated
using the shortest ellipsoidal path (great circle on a sphere).}
}
\value{
A \code{\link[=carrow_schema]{carrow_schema()}}.
}
\description{
These schemas are used as the basis for column types in Apache Arrow
}
\examples{
geo_arrow_schema_point()
geo_arrow_schema_linestring()
geo_arrow_schema_polygon()
geo_arrow_schema_multi(geo_arrow_schema_point())
geo_arrow_schema_sparse_geometrycollection()
geo_arrow_schema_dense_geometrycollection()

}
