% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoarrow-package.R
\docType{package}
\name{geoarrow-package}
\alias{geoarrow}
\alias{geoarrow-package}
\title{geoarrow: Extension types for geospatial data for use with 'Arrow'}
\description{
Provides extension types and conversions to between R-native object types and 'Arrow' columnar types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geoarrow/geoarrow-c}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
