% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{geoarrow}
\alias{geoarrow}
\alias{geoarrow_wkb}
\alias{geoarrow_wkt}
\title{Create GeoArrow Vectors}
\usage{
geoarrow(x = wk::wkb())

geoarrow_wkb(x = wk::wkb())

geoarrow_wkt(x = wk::wkb())
}
\arguments{
\item{x}{An object that implements \code{\link[wk:wk_handle]{wk::wk_handle()}}}
}
\value{
A vector class inheriting from 'geoarrow_vctr'.
}
\description{
Create GeoArrow Vectors
}
