% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{geoarrow_create}
\alias{geoarrow_create}
\alias{geoarrow_create.default}
\alias{geoarrow_schema_default}
\title{Create GeoArrow Arrays}
\usage{
geoarrow_create(
  handleable,
  ...,
  schema = geoarrow_schema_default(handleable),
  strict = FALSE
)

\method{geoarrow_create}{default}(
  handleable,
  ...,
  schema = geoarrow_schema_default(handleable),
  strict = FALSE
)

geoarrow_schema_default(
  handleable,
  point = geoarrow_schema_point(nullable = FALSE)
)
}
\arguments{
\item{handleable}{An object with a \code{\link[wk:wk_handle]{wk::wk_handle()}} method}

\item{...}{Passed to \code{\link[wk:wk_handle]{wk::wk_handle()}}}

\item{schema}{A \code{\link[narrow:narrow_schema]{narrow::narrow_schema()}} to use as a storage method.}

\item{strict}{Use \code{TRUE} to respect choices of storage type, dimensions,
and CRS provided by \code{schema}. The default, \code{FALSE}, updates these values
to match the data.}

\item{point}{The point schema to use for coordinates}
}
\value{
A \code{\link[narrow:narrow_array]{narrow::narrow_array()}}
}
\description{
Create GeoArrow Arrays
}
\examples{
geoarrow_create(wk::xy(1:5, 1:5))

}
