% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-geoarrow.R
\name{wk_handle.carrow_array}
\alias{wk_handle.carrow_array}
\alias{wk_handle.carrow_array_stream}
\title{Handle Arrow arrays}
\usage{
\method{wk_handle}{carrow_array}(handleable, handler, ...)

\method{wk_handle}{carrow_array_stream}(
  handleable,
  handler,
  ...,
  geoarrow_schema = carrow::carrow_array_stream_get_schema(handleable),
  geoarrow_n_features = NA_integer_
)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}

\item{handler}{A \link[wk:wk_handle]{wk_handler} object.}

\item{...}{Passed to the \code{\link[wk:wk_handle]{wk_handle()}} method.}
}
\value{
The result of \code{handler}
}
\description{
Handle Arrow arrays
}
